Fix checks not stopping the command
