+++ src/UrfRidersBot.Infrastructure/Commands/CommandContext.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Text;
 using System.Threading.Tasks;
 using DSharpPlus;
 using DSharpPlus.Entities;
+using UrfRidersBot.Core;
 using UrfRidersBot.Core.Commands;
 
 namespace UrfRidersBot.Infrastructure.Commands
@@ -72,6 +74,40 @@ namespace UrfRidersBot.Infrastructure.Commands
                 .AddEmbed(embed);
 
             return await RespondAsync(builder);
+        }
+
+        public async ValueTask<DiscordMessage> RespondWithCriticalErrorAsync(Exception exception)
+        {
+            if (IsEphemeral)
+            {
+                var sb = new StringBuilder();
+                sb.AppendLine($"{UrfRidersEmotes.HighPriority} Command failed, please contact bot owner.");
+                sb.AppendLine($"Exception: {Markdown.Code(exception.Message)}");
+
+                return await RespondAsync(sb.ToString());
+            }
+
+            var embed = EmbedHelper.CreateCriticalError("Please contact bot owner.", "Command failed");
+            embed.AddField("Exception", Markdown.Code(exception.Message));
+
+            return await RespondAsync(embed);
+        }
+
+        public async ValueTask<DiscordMessage> RespondWithAccessDeniedAsync(string? reason = null)
+        {
+            reason ??= "You don't have permission to execute this command.";
+            
+            if (IsEphemeral)
+            {
+                var sb = new StringBuilder();
+                sb.AppendLine($"{UrfRidersEmotes.Unavailable} Access Denied");
+                sb.AppendLine(reason);
+            
+                return await RespondAsync(sb.ToString());
+            }
+            
+            var embed = EmbedHelper.CreateUnavailable(reason, "Access Denied");
+            return await RespondAsync(embed);
         }
     }
 }
